Fix git/npm ignores
Compiled distributables shouldn't be in git (one doesn't version
compiled output) so this removes dist from the repo and adds it to
.gitignore.

npm will use .gitignore *as* .npmignore if npmignore is not present. It
is preferable to leverage this behavior so that ignore rules don't need
duplicated between two files; which is error prone and will certainly
lead to these files drifting unententionally. Further, the nature of
.npmignore means that any errornous ignores are very unlikely to get
caught until after a release has been published to the registry. That is
undesirable.