npm adds node_modules/.bin to PATH
Again, same issue as https://github.com/optimizely/javascript-sdk/pull/56, there is no need to point directly at the node_modules bin because npm purposefully prepends `$(npm bin)` to `PATH` when running scripts.

(Which is in addition to the fact that `./node_modules/karma/bin/karma` is the wrong path. It's brittle and reliant on karma always keeping that file in the same location with the same name. Correct location for binaries is where npm symlinks them, as instructed by the `bin` property of package.json: `./node_modules/.bin/karma`. But again, all is unnecessary because npm has the feature specifically to avoid this.)