fix (typed audience evaluation): When evaluating exact match condition, properly check type of condition value (#178)
Fixes a bug in evaluation of exact match conditions. Before this change, evaluation of a condition with a wrongly-typed value property could return a boolean, when it should abort and return null. With this change, when condition.value of an exact match condition is not a string, boolean, or finite number, the condition will evaluate to null.