refactor (datafile manager): Update Headers type to only store at-most 1 value per header name (#5)
Summary:

Change `Headers` type to have at-most 1 value per header name. This is sufficient for the current requirements (which only use If-Modified-Since and Last-Modified), and simplifies the code. We can introduce a richer type if we need to deal with multiple values for headers.

Test plan:

Existing unit tests. This should be a pure refactor with no behavior change.