feat (datafile manager): Increasing backoff delay after network errors or non-success response status (#7)
Summary:

Implement an increasing delay for the next request after a response Promise is rejected, or resolved with a non-successful status.

If the configured updateInterval is greater than the calculated delay, we just use the updateInterval.

Test plan:

Added unit tests for new module backoffController. Added unit tests for httpPollingDatafileManager backoff functionality.