fix (datafile manager): Change onReady meaning to 'a datafile is available' (#14)
Summary:

This changes the behavior of the `onReady` Promise to align with the Java implementation. The `onReady` Promise is now resolved as soon as the datafile manager has any datafile (whether it was passed in or fetched). An update event is always emitted when the datafile changes after `onReady`.

Test plan:

Updated unit tests