fix (unit tests): Fix event dispatcher missing in unit test (#256)
Summary:

Optimizely was constructed without an event dispatcher, causing an error to be thrown asynchronously. This fixes the test, and updates the npm test command to make it easier to catch such mistakes - adding a listener for 'unhandledRejection' on process that calls process.exit.

Test plan:

Run unit tests, verify log output.