fix (datafile management): Clear timeout created in onReady method when it's no longer needed (#272)
Summary:

onReady sets a timeout that fulfills a timeout promise. The timeout promise is only used to race against the ready promise provided by ProjectConfigManager. So, as soon as ProjectConfigManager's ready promise is fulfilled, the timeout can be cleared. Before this change, the timeout would only be cleared in the close method.

Test plan:

Manual & added unit test