feat(event processor)[OASIS-4921] Return Promise from close method representing the close process (#295)
Summary:

The close method flushes the event processor's queue. If the queue was non-empty, requests will be sent to the event dispatcher. Users may want to wait until the dispatcher finishes sending these, so we return a Promise that fulfills after the event dispatcher calls the "done" callback for each request.

The event processor already returns a Promise from its stop method, so this change is just making use of that.

Test plan:

New unit tests. Manually tested