fix(event processor): Flush current batch when event with different context received (#341)
Summary:

Prior to this change, event processor's buffer could contain events with different contexts. At flush time, the buffered events would be grouped by context and a request would be dispatched for each group.

With this change, we introduce an invariant: all events in the buffer share the same context. As soon as an event with different context is received, the queue is flushed. With this restriction, at most one request is dispatched per flush.

- A new comparator function (batchComparator) is provided to DefaultEventQueue. When enqueuing an event, batchComparator is used to determine whether the incoming event can be included into the current batch. If not, we flush.
- LogTierV1EventProcessor provides a batchComparator when constructing a DefaultEventQueue that checks equality of all context properties.
- In AbstractEventProcessor, the grouping step is removed from the flush process.

Test plan:

New unit tests. Manually tested.

Issues:

https://optimizely.atlassian.net/browse/OASIS-5160