feat (event processor/pending events dispatcher): In browser entry point, add event listener for pagehide or unload that calls optimizely.close (#347)
Summary:

With event batching, optimizely.close() needs to be called to ensure that events are flushed before the user navigates away.  This PR ensures that optimizely.close() is called when the page unloads by hooking into the 'pagehide' event or the 'unload' event.

Test plan:

Added a UMD bundle test. Manually tested.