feat (Optimizely) [OASIS-5907]: close promise tracks in-flight event dispatcher requests (#404)
Summary:

Update to js-sdk-event-processor v0.4.0, which supports tracking in-flight event dispatcher requests in its stop() promise. The close promise (of Optimizely) is based on the event processor's stop promise, so by this upgrade, we get the new functionality. I updated the close doc comment to talk about this new behavior.

Test plan:

- Existing tests
- Manually tested in Node, browser, and Lambda environments

Issues:

https://optimizely.atlassian.net/browse/OASIS-5907