feat(core)[OASIS-6054]: Change Error to LOG_LEVEL.WARNING for unfound event
In a similar vein as https://github.com/optimizely/java-sdk/pull/361,
the JavaScript SDK would be less noisy when being used with integrations
like Segment if we only logged a warning when an event key is not found
in the datafile. This PR will make that change and unblock event
tracking.