chore: Update dev dependencies to prevent fsevents install error message (#433)
Summary:

Optional transitive dev dependency fsevents fails to install using Node 12.16.1, and possibly other versions. fsevents is pulled in by webpack and karma.

In this PR I updated karma and webpack to the latest versions, which depend on newer versions of fsevents, for which the error doesn't occur. I also updated karma-webpack to the latest version.

Test plan:

- Ran karma tests (xbrowser, umd)
- Inspected UMD build output with newer webpack version

Issues:

https://github.com/optimizely/javascript-sdk/issues/432