Use rollup's --config- option for choosing bundles
Rollup provides a feature specifically designed for passing in config
options to the rollup configuration module. That is: --config-* options.
(It is documented deeply under
https://rollupjs.org/guide/en/#configuration-files )

I believe this is a superior mechanism for passing options to rollup's
config because they do not impact process.env.

Environment variables have a very specific capability, which is that
they are inherited by child processes. These are settings that quite
literally "configure the environment under which a process runs".

However, their use prior to this commit was an abuse in order to simply
pass an _option_ to the rollup configuration that would control which
bundles to build. Rollup's --config* feature is superior for this role,
because the option is only intended to go so far as the configuration
module itself. Thus, using them to control which bundle(s) are built
does not leak extraneous information into the *ahem* environment of the
build process.

As a side benefit, the cli invocation is cleaner!