Fix packages for use as git-deps and npm-linking
`prepublishOnly` is an old lifecycle hook created to aid in the
transition from prepublish to prepare.

Documentation: https://docs.npmjs.com/misc/scripts#prepublish-and-prepare

TL;DR: prepare runs before packing and on bare npm-installs.
That is, in order for a package to do building/compiling when it is
installed as a git-dep (where prepublish _does not run_), one must use
`prepare`. This also applies for when a package is symlinked using
npm-link. In order to ensure the package is built and can be consumed by
the linked dependant, the build must occur via `prepare`.

It's also worth noting that `prepare` runs _before_ `prepublish`; so the
build step is guaranteed to be complete prior to the prepublish script;
for, say, running tests.