Ignore npm-pack'd tarballs
It's frequently necessary to run npm-pack when validating the package
tarball is created correctly. This ensures the tarballs are themselves
ignored from git.

(Keeping user-specific ignores, but those shouldn't be in a repo ignore.
User-specific ignores that are OS or IDE specific should be in the
user's own ignore file.)