Fix broken test suite
XMLHttpRequest is stubbed by sinon but never assigned to the global!
Thus, the stubbing is never used anywhere in this test suite _except_
that it relies on the stubbing done in _other_ test files that aren't
cleaned up.

This ensures the stubbing is set and used within this test file. (But
does not clean up the bad global stubbings that are left intact after
each test run.)

This partially fixes https://github.com/optimizely/javascript-sdk/issues/488