Avoid cross-test closure variables
Any variable that is declared outside of a test will leak state between
tests. This is bad and dangerous. Variables that are necessary to
reference between both setup and test callbacks should be defined on the
test context, which mocha creates fresh for each test. This way,
cross-test pollution is minimized. (This is part of the problem that
caused: https://github.com/optimizely/javascript-sdk/issues/488)