fixes the `e.replace is not a function` error happening on production in Firefox during createInstance
in js, any object can be thrown. that is why the `catch` block `error` variable has type `unknown` in typescript. in optimizely, there are a few catch blocks where the error variable is directly logged as message. in the logger, the type of the public named logging functions is `string | Error` which is not correct since it could actually be any type. if it is an object that is not an instance of type Error, this will cause a throw in `internalLog -> format -> sprintf` because sprintf is calling `.replace` on the message, which throws with `e.replace is not a function`. we have seen this exact error a lot in production. so logging causes a throw which causes createInstance to fail, causing further issues.

this PR fixes the issue by converting non-string and non-Error log messages to string first.

fixes #719