revert #735 and prep for release 0.9.3 for `js-sdk-datafile-manager` and 0.9.4 for `js-sdk-event-processor` (#739)
## Summary
React native dependencies were converted from `peer` to `optional` in `js-sd-datafile-manager` version `0.9.2` and `js-sdk-event-processor` version `0.9.3`. After releasing, it appeared that they did not work as expected. This PR reverts back that change which was made in #735 and updates changelog and npm version to make new releases for both the packages.

## Test plan
- Manually tested.
- All unit tests pass