[FSSDK-9611] fix relative import paths that goes out of lib (#856)
we are bundling our package, and type definiton of all our dependencies are in dev-dependencies as they should be and therefore not installed when clients install our sdk.compiling our lib files will need those type definitions. so files in our dist must not refer to the lib files, but some of them do refer to the lib files. and consumers are getting missing types as a result.

files in /dist was referring to files in /lib because some relative import paths
were going out of lib and then back in. as the path strings were
preserved in the built files in dist, it was going out of /dist and into
/lib.

This commit fixes those relative import paths to stay inside /lib