[FSSDK-11497] fire config update events synchronously (#1043)
fire the project config updates synchronously. this will allow the optimizely class to propagate the initial project config synchronously to all sub components. This can be useful when the client is initialized with a datafile. One downside is, the update event handler needs to be registered before calling start() on the config manager, but since we only use the config manager internally, we can maintain that.