[FSSDK-11898] serialize concurrent cmab service calls (#1086)
The cmab service caches the results of a cmab prediction retrieve
from the server and returns it for subsequent call. This ensures
a consistent value is returned for getDecision() within the cache ttl.
However, when there is no cached value, if there is concurrent calls
to gertDecision() for same userId and ruleId combination, all of these
will cause a call to the server and may potentially return different
values.

The solution is to run concurrent calls for same userId and ruleId
combinations one after another. To achieve this, we put each (userId, ruleId)
combination in one of the predefined bucktes by hashing the
(userId, ruleId) combination and serialize all calls for that
particular hash % (num_buckets).