[FSSDK-11879] flush events without closing client on page unload (#1087)
Currently, in browser, when the page is unloaded, the sdk instance is being closed in order to flush pending events. But when the page is loaded from [bfcache](https://developer.mozilla.org/en-US/docs/Glossary/bfcache), the sdk instance stays closed as bfcache restores the Javascript heap as well, which causes further events to be not processed. 

This PR updates the code to flush events on page unload without closing the sdk instance.