[FSSDK-12028] add ttl to events in event store (#1103)
Add configurable ttl support to stored events. This will be helpful if for some reason, events are failing permanently (for example, the event enpoint is blocked by adblocker/isp).

Also, refactored the code to move the responsibility of limiting the number of saved event to the store implementation from the event processor.